Module liq_queue::bid

Functions

assert_bid_asset_from_sent_funds

Validate sent assets

calculate_liquidated_collateral

Calculate the amount of collateral to liquidate

calculate_remaining_bid

Calculate the remaining bid amount after a scale change, i.e. a liquidation or a bid activation

claim_liquidations

Bid owner can claim their share of the liquidated collateral until the bid is consumed

execute_liquidation

Positions contract (owner) executes the liquidation and pays in the msg reply This operation returns a repay_amount based on the available bids on each premium slot, consuming bids from lowest to higher premium slots

read_bid

Return Bid from storage

read_bids_by_user

Return Bids for a user

read_epoch_scale_sum

Read epoch scale sum

read_premium_slot

Read premium slot

retract_bid

Withdraw bid amount

store_epoch_scale_sum

Store epoch scale sum

submit_bid

Create Bid and add to the corresponding Slot

validate_bid_input

Validate bid input